requirements.txt
crewai
langchain
langchain-google-genai
requests
python-dotenv
crewai-tools
